BOCM-80 - moves getAppealReview controller tests to own file
